requirements.txt
click==8.0.1
cycler==0.10.0
importlib-metadata==4.6.3
itsdangerous==2.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
opencv-python==4.5.3.56
pandas==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
